[DDW-1197] Move ExperimentalDataOverlay component to notifications folder
