+++ source/renderer/app/components/notifications/ExperimentalDataOverlay.js
@@ -4,7 +4,7 @@ import { defineMessages, intlShape, FormattedHTMLMessage } from 'react-intl';
 import SVGInline from 'react-svg-inline';
 import { Button } from 'react-polymorph/lib/components/Button';
 import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
-import testnetBulbIcon from '../../../assets/images/testnet-bulb.inline.svg';
+import testnetBulbIcon from '../../assets/images/testnet-bulb.inline.svg';
 
 import styles from './ExperimentalDataOverlay.scss';
 
@@ -41,9 +41,7 @@ export default class ExperimentalDataOverlay extends Component<Props> {
     const { onClose } = this.props;
     const title = intl.formatMessage(messages.title);
     const actionLabel = intl.formatMessage(messages.actionLabel);
-    const description = (
-      <FormattedHTMLMessage {...messages.description} />
-    );
+    const description = <FormattedHTMLMessage {...messages.description} />;
 
     return (
       <div className={styles.component}>
