+++ source/renderer/app/components/notifications/ExperimentalDataOverlay.scss
@@ -64,9 +64,9 @@
             );
             content: '';
             height: 14.5px;
-            -webkit-mask: url('../../../assets/images/experiment-icon.svg')
+            -webkit-mask: url('../../assets/images/experiment-icon.svg')
               no-repeat center;
-            mask: url('../../../assets/images/experiment-icon.svg') no-repeat
+            mask: url('../../assets/images/experiment-icon.svg') no-repeat
               center;
             -webkit-mask-size: 10px;
             mask-size: 10px;
